Add host, port, reload and changed the secret_key value from the .env Add all user functions are added in AccessUsers class. Removed the initializing of settings and get it directly from config.py. Fixed the imports and removed all the user functions and moved it to user_dependencies.py. Removed all the routes and moved the token route to user_token.py.
