+++ api/models/models.py
@@ -4,7 +4,6 @@ from api.database import Base, engine
 
 
 class User(Base):
-    __table_args__ = {'extend_existing': True}
     __tablename__ = "registered_users"
 
     id = Column(Integer, primary_key=True)
