+++ requirements.txt
@@ -9,3 +9,4 @@ cryptography==41.0.7
 sqlalchemy==2.0.23
 pydantic-settings==2.1.0
 pydantic==2.5.2
+requests==2.31.0
